Quite more 64-to-32-bit warnings.

+++ SolitaireBakersGame.m
@@ -44,7 +44,7 @@
     
     int i;
     for(i = 0; i < 8; i++) {
-        int pos = [tableau_[i] count] - 2;
+        NSInteger pos = [tableau_[i] count] - 2;
         while(pos >= 0) {
             SolitaireCard* card = [tableau_[i] cardAtPosition: pos];
             if(([card.nextCard faceValue] != [card faceValue] - 1) ||
@@ -72,7 +72,7 @@
 -(void) dropCard: (SolitaireCard*) card inTableau: (SolitaireTableau*) tableau {
     [super dropCard: card inTableau: tableau];
         
-    int pos = [tableau count] - 2;
+    NSInteger pos = [tableau count] - 2;
     while(pos >= 0) {
         SolitaireCard* card = [tableau cardAtPosition: pos];
         if(([card.nextCard faceValue] == [card faceValue] - 1) &&
@@ -83,7 +83,7 @@
 }
 
 -(void) onCard: (SolitaireCard*) card removedFromTableau: (SolitaireTableau*) tableau {
-    int pos = [tableau count] - 2;
+    NSInteger pos = [tableau count] - 2;
     while(pos >= 0) {
         SolitaireCard* card = [tableau cardAtPosition: pos];
         if(([card.nextCard faceValue] == [card faceValue] - 1) &&
